+++ grunt/ci-pull.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('ci-pull', ['test', 'jasmine:specs:build', 'connect:server']);
+};
