+++ grunt/ci.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('ci', ['test', 'jasmine:specs:build', 'connect:server', 'saucelabs-jasmine']);
+};
