+++ grunt/coverage.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('coverage', ['build', 'jasmine:coverage']);
+};
