+++ grunt/docs.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('docs', ['build', 'copy', 'emu', 'toc', 'markdown', 'docco']);
+};
