+++ grunt/lint.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('lint', ['build', 'jshint', 'jscs']);
+};
