+++ grunt/merge.js
@@ -1,0 +1,6 @@
+module.exports = function(grunt) {
+	grunt.registerTask('merge', 'Merge a github pull request.', function (pr) {
+        grunt.log.writeln('Merge Github Pull Request #' + pr);
+        grunt.task.run(['shell:merge:' + pr, 'test' , 'shell:amend']);
+    });
+};
