+++ grunt/options/browserify.js
@@ -1,0 +1,11 @@
+module.exports = {
+	dev: {
+        src: '<%= conf.pkg.name %>.js',
+        dest: 'bundle.js',
+        options: {
+            browserifyOptions: {
+                standalone: 'dc'
+            }
+        }
+    }
+};
