+++ grunt/options/concat.js
@@ -1,0 +1,11 @@
+module.exports = {
+    options: {
+        process: true,
+        sourceMap: true,
+        banner : '<%= conf.banner %>'
+    },
+    js: {
+        src: '<%= conf.jsFiles %>',
+        dest: '<%= conf.pkg.name %>.js'
+    }
+};
