+++ grunt/options/connect.js
@@ -1,0 +1,8 @@
+module.exports = {
+	server: {
+        options: {
+            port: 8888,
+            base: '.'
+        }
+    }
+};
