+++ grunt/options/docco.js
@@ -1,0 +1,13 @@
+module.exports = {
+	options: {
+        dst: '<%= conf.web %>/docs',
+        basepath:'<%= conf.web %>'
+    },
+    howto: {
+        files: [
+            {
+                src: ['<%= conf.web %>/stock.js']
+            }
+        ]
+    }
+};
