+++ grunt/options/emu.js
@@ -1,0 +1,6 @@
+module.exports = {
+	api: {
+        src: '<%= conf.pkg.name %>.js',
+        dest: '<%= conf.web %>/docs/api-latest.md'
+    }
+};
