+++ grunt/options/gh-pages.js
@@ -1,0 +1,7 @@
+module.exports = {
+	options: {
+        base: '<%= conf.web %>',
+        message: 'Synced from from master branch.'
+    },
+    src: ['**']
+};
