+++ grunt/options/markdown.js
@@ -1,0 +1,7 @@
+module.exports = {
+	html: {
+        src: '<%= emu.api.dest %>',
+        dest: '<%= conf.web %>/docs/index.html'
+    },
+    options: {markdownOptions: {highlight: 'manual'}}
+};
