+++ grunt/options/toc.js
@@ -1,0 +1,6 @@
+module.exports = {
+	api: {
+        src: '<%= emu.api.dest %>',
+        dest: '<%= emu.api.dest %>'
+    }
+};
