+++ grunt/server.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('server', ['docs', 'fileindex', 'jasmine:specs:build', 'connect:server', 'watch:jasmine']);
+};
