+++ grunt/test-browserify.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('test-browserify', ['build', 'browserify', 'jasmine:browserify']);
+};
