+++ grunt/test-stock-example.js
@@ -1,0 +1,6 @@
+module.exports = function(grunt) {
+	grunt.registerTask('test-stock-example', 'Test a new rendering of the stock example web page against a ' +
+        'baseline rendering', function (option) {
+            require('./regression/stock-regression-test.js').testStockExample(this.async(), option === 'diff');
+        });
+};
