+++ grunt/test.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('test', ['build', 'jasmine:specs', 'shell:hooks']);
+};
