+++ grunt/update-stock-example.js
@@ -1,0 +1,5 @@
+module.exports = function(grunt) {
+	grunt.registerTask('update-stock-example', 'Update the baseline stock example web page.', function () {
+        require('./regression/stock-regression-test.js').updateStockExample(this.async());
+    });
+};
