+++ grunt/web.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+	grunt.registerTask('web', ['docs', 'gh-pages']);
+};
