+++ package.json
@@ -30,6 +30,7 @@
   },
   "devDependencies": {
     "emu": "~0.0.2",
+    "glob": "^5.0.5",
     "grunt": "~0.4.5",
     "grunt-browserify": "~3.3.0",
     "grunt-cli": "~0.1.13",
@@ -50,11 +51,11 @@
     "grunt-saucelabs": "~8.5.0",
     "grunt-shell": "~1.1.1",
     "grunt-template-jasmine-istanbul": "~0.3.2",
-    "load-grunt-tasks": "~3.0.0",
     "jsdifflib": "~1.1.0",
+    "load-grunt-tasks": "~3.0.0",
     "marked": "~0.3.3",
-    "uglify-js": "2.4.x",
-    "time-grunt": "~1.0.0"
+    "time-grunt": "~1.0.0",
+    "uglify-js": "2.4.x"
   },
   "scripts": {
     "test": "grunt test"
